Extract serializeArgsForDependency into a shared module
The helper that collapses query arguments into a stable string dependency
lived privately inside useObservableQuery, so only the observable hooks
benefited from it. Moving it next to its siblings lets every query hook
answer "how are arguments compared" the same way, which is the actual
seam the useQuery loop exposed.

No behavior change - useObservableQuery keeps calling the same function.
